refactor(omp): split rpc_bridge modules for file-length gate
